feat: support apollo-plugin
This is almost a rewrite, so there might be bugs lurking.
I dropped the unit tests as they were relatively high maintenance for the value they provide.
To replace them I created integration tests for every major use-case.
Having snapshots of the span tree allows us to visualize these tests well enough so that
one can expect outside contributors to be able to participate here.

BREAKING CHANGE: This removes support for the GraphQL Extension interface and moves from a class to a function interface